TwitterBot/requirements.txt
tweepy
cairosvg
sseclient
python-dotenv
pickle-mixin
